ID-B-9.3-move-shared-Dtos
