--- backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Commands/AddPet/AddressDto.cs
@@ -1,3 +1,0 @@
-namespace FindYourFriendAmongPets.Application.Volunteers.Commands.AddPet;
-
-public record AddressDto(string City, string Street, string? Building = null, string? Description = null, string? Country = null);
