--- backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Commands/AddPet/PetSpeciesDto.cs
@@ -1,3 +1,0 @@
-namespace FindYourFriendAmongPets.Application.Volunteers.Commands.AddPet;
-
-public record PetSpeciesDto(Guid SpeciesId, Guid BreedId);
