--- backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Commands/Create/RequisiteForHelpDto.cs
@@ -1,5 +1,0 @@
-﻿namespace FindYourFriendAmongPets.Application.Volunteers.Commands.Create;
-
-public record RequisiteForHelpDto(
-    string Name,
-    string Description);
