--- backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Commands/Create/SocialNetworkDto.cs
@@ -1,5 +1,0 @@
-﻿namespace FindYourFriendAmongPets.Application.Volunteers.Commands.Create;
-
-public record SocialNetworkDto(
-    string Title,
-    string Link);
