+++ backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Shared/AddressDto.cs
@@ -1,0 +1,8 @@
+namespace FindYourFriendAmongPets.Application.Volunteers.Shared;
+
+public record AddressDto(
+    string City, 
+    string Street, 
+    string? Building = null, 
+    string? Description = null, 
+    string? Country = null);
