+++ backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Shared/PetSpeciesDto.cs
@@ -1,0 +1,3 @@
+namespace FindYourFriendAmongPets.Application.Volunteers.Shared;
+
+public record PetSpeciesDto(Guid SpeciesId, Guid BreedId);
