+++ backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Shared/RequisiteForHelpDto.cs
@@ -1,0 +1,6 @@
+﻿namespace FindYourFriendAmongPets.Application.Volunteers.Shared;
+
+public record RequisiteForHelpDto(
+    string Name,
+    string Description,
+    Guid? VolunteerId);
