+++ backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Shared/SocialNetworkDto.cs
@@ -1,0 +1,6 @@
+﻿namespace FindYourFriendAmongPets.Application.Volunteers.Shared;
+
+public record SocialNetworkDto(
+    string Title,
+    string Link,
+    Guid? VolunteerId);
